refactor: migrate tests from test/ to test-env/edge-cases/
Consolidated test infrastructure by migrating legacy rollback tests
to the modern test-env structure.

Migration Summary:
- Moved 3 rollback test files to test-env/edge-cases/
- Converted from custom runTest pattern to node:test framework
- Updated to use node:assert/strict
- Removed console.log and manual test counting
- Organized tests with proper describe/test structure

Migrated Tests (69 tests total):
- rollback-edge-cases.test.js (58 tests): Comprehensive security
  validation including shell injection, path traversal, hash validation
- rollback-user-sections.test.js (4 tests): USER section extraction
- rollback-validation.test.js (7 tests): GREEN phase rollback validation

Deleted Legacy Files:
- test/rollback-*.test.js
- test/agents-md/, test/plugins/, test/validation/ subdirectories
- test/TEST_COVERAGE.md

Benefits:
- Single unified test infrastructure (test-env/)
- Modern node:test framework throughout
- Consistent test patterns and structure
- Total: 189 tests passing (120 Phase 3 + 69 rollback)

All tests verified passing before deletion.
